refactor(runtime): route loop_::run tool assembly through the ScopedToolRegistry seam
loop_::run hand-rolled the per-agent tool registry (peripherals -> built-in
allowed_tools/excluded_tools filter -> MCP scope+gate -> skill registration) inline,
one of the sites the assemble() seam (Epic A, #8640) exists to collapse. Route it
through scoped::ScopedToolRegistry::assemble.

Behavior-neutral - verified two ways:
* The tool REGISTRY is identical: same tools, same order (peripherals -> filter ->
  MCP capability tools -> pinned -> deferred/eager -> skills), same policy gating.
  run() used the plain apply_policy_tool_filter with its per-run allowed_tools, which
  maps 1:1 onto assemble's caller_allowed; peripherals (load_peripheral_tools IS the
  PERIPHERAL_TOOLS_FN wrapper run() called), the MCP step, and skills are what run()
  hand-rolled. Confirmed by an independent multi-agent diff and by a separate Codex
  review (both: NEUTRAL, nothing beyond logs).
* The per-step diagnostic INFO logs are PRESERVED, not dropped. They are folded into
  the seam behind a new  knob: the peripheral count, the
  'Applied capability-based tool access filter' before/after audit line, and the MCP
  init/deferred/eager count logs now emit from assemble() when the flag is set.
  run() passes true (keeps its exact breadcrumbs); the gateway's two listing builders
  pass false so a registry no turn runs against does not emit spurious 'MCP: N
  registered' / 'Peripheral tools added' lines. Future execution paths get the
  diagnostics for free - this is the seam-owned observability parity the program
  wants, delivered here rather than deferred.

The only observable delta: those five diagnostic logs now carry module=tools::scoped
(where assembly happens) instead of agent::loop_; message, attrs, category, and action
are identical. The gateway stays exactly as it was (flag=false, no logs).

Scope: loop_::run + the seam-diagnostics knob + the gateway flag wiring. process_message
is a separate follow-up (it filters via filter_channel_builtin_tools, the safe-defaults
admit - its own behavior-flagged PR).

Validation (toolchain 1.93.0): clippy -p zeroclaw-runtime -p zeroclaw-gateway
-p zeroclaw-channels --all-targets -- -D warnings clean; runtime lib 2756 passed;
gateway lib 332 passed; fmt clean.
